+++ src/lib/hooks/index.ts
@@ -1,3 +1,4 @@
 export * from './useAlert';
+export * from './useAxiosErrorAlert';
 export * from './useConfirm';
 export * from './useModal';
